feat: add Redis connection resilience and production testing
## Connection Management
- Add configurable timeouts (connection: 30s, response: 20s)
- Integrate ConnectionManager with timeout configuration
- Implement persistent retry with exponential backoff (1s→60s)
- Never give up reconnecting to Redis (infinite retry)

## PubSub Auto-Reconnect
- Add automatic reconnection after connection loss
- Auto re-subscribe to all channels after reconnect
- 5-second delay between reconnect attempts
- Non-blocking background task for subscribe operations

## Code Quality
- Apply all Clippy fixes (unnecessary casts, needless borrows)
- Remove unused code where appropriate
- All public APIs maintained (zero breaking changes)

## Documentation
- Add CHANGELOG.md with version 1.0.1 release notes
- Add PRODUCTION_DEPLOYMENT.md with provider configurations
- Add PUBSUB_AUTORECONNECT.md with feature documentation
- Update README.md with new features and tested environments
- Update CONNECTION_SUPERVISION.md with timeout configurations
- Update MONITORING.md with production resilience section
- Update IMPLEMENTATION_COMPLETE.md with recent changes

## Tested Platforms
- ✅ Self-hosted Redis (local/Docker/K8s)
- ✅ Upstash (Redis Cloud)
- ✅ Aiven Valkey (Redis-compatible)

Fixes:
- Resolve "timed out" errors with cloud-hosted Redis
- Fix PubSub consumers not receiving messages after Redis restart
- Fix connection hangs with proper timeout configuration

All multi-consumer logic remains unchanged (100% backward compatible)

+++ src/retry.rs
@@ -153,7 +153,7 @@ where
                 if is_connection_error && attempt < config.max_attempts {
                     // Add jitter to prevent thundering herd
                     let jitter = if config.jitter_ms > 0 {
-                        rand::random::<u64>() % config.jitter_ms as u64
+                        rand::random::<u64>() % config.jitter_ms
                     } else {
                         0
                     };
